refactor(api): split exec handler responsibilities
